fixed some PR feedback, moved some things to internal
