+++ cli/README.md
@@ -43,7 +43,7 @@ $ ssv-dkg sign --input /path/to/deposit/data \
 ```
 Each operator must be in the form `$validatorNonce,$address`. Providing the wrong validator nonce may result in disaster for your DKG.
 The output directory will default to `~/.ssv`. It will be in a file named after the date (and a counter if you create multiple clusters in a day). 
-You will need to maintain this state file if you wish to reshare the key for this cluster in future, e.g. if operators become unresponsive and you wish to exclude them. 
+You will need to maintain this state file if you wish to reshare the key for this cluster in the future, e.g. if operators become unresponsive and you wish to exclude them. 
 
 - combine both in a single command
 ```shell
@@ -61,4 +61,4 @@ $ ssv-dkg reshare --state ~/.ssv/deadbeefcafebabe.json \
 ⏳ starting distributed key resharing
 ✅ distributed key reshared successfully!
 ```
-Note: you will have to maintain 50% + 1 of the same operators in the new cluster for this to succeed.
+Note: you will have to maintain a majority of operators from one cluster to the next.
