+++ internal/dkg_error_processes.go
@@ -1,4 +1,4 @@
-package integration
+package internal
 
 import (
 	"errors"
