+++ internal/integration_test.go
@@ -1,4 +1,4 @@
-package integration
+package internal
 
 import (
 	"fmt"
